Safari-611.1.10/LayoutTests/editing/deleting/smart-delete-paragraph-004.html

<!DOCTYPE html>
<html> 
<head>
<style>
.editing { 
    border: 2px solid red; 
    font-size: 24px; 
}
.explanation { 
    border: 2px solid blue; 
    padding: 12px; 
    font-size: 24px; 
    margin-bottom: 24px;
}
.scenario { margin-bottom: 16px;}
.scenario:first-line { font-weight: bold; margin-bottom: 16px;}
.expected-results:first-line { font-weight: bold }
</style>
</head> 
<body>
<div class="explanation">
<div class="scenario">
Tests: 
<br>
Smart delete when removing the last paragraph.
</div>
<div class="expected-results">
Expected Results:
<br>
Extra newlines should be removed to maintin spacing between paragraphs. It should like this:<BR>
Test paragraph.<BR>
<br>
Middle test paragraph.<br>
</div>
</div>

<div contenteditable id="root" class="editing" style="word-wrap: break-word; -khtml-nbsp-mode: space; -khtml-line-break: after-white-space;">
<div>
Test paragraph.
</div>
<div><br></div>
<div>
Middle test paragraph.
</div>
<div><br></div>
<div id="test">
Last test paragraph to remove.
</div>
</div>
<script src="../editing.js"></script>
<script src="../../resources/ui-helper.js"></script>
<script>

async function editingTest() {
    internals.settings.setEditingBehavior("ios");
    await UIHelper.selectWordByDoubleTapOrClick(document.getElementById('test'));
    extendSelectionForwardByLineBoundaryCommand();
    cutCommand();
}

runDumpAsTextEditingTest(true);
</script>
</body>
</html>
